Move uts which startup ldap servers to its.
