+++ nexus/plugins/ldap/nexus-ldap-realm-plugin/src/test/java/org/sonatype/nexus/LdapNexusIT.java
@@ -12,20 +12,16 @@
  */
 package org.sonatype.nexus;
 
-import java.io.FileOutputStream;
-import java.io.IOException;
-
 import org.apache.shiro.authc.UsernamePasswordToken;
 import org.apache.shiro.subject.SimplePrincipalCollection;
-import org.codehaus.plexus.util.IOUtil;
 import org.junit.Assert;
 import org.junit.Test;
 import org.sonatype.nexus.security.ldap.realms.NexusLdapAuthenticationRealm;
 import org.sonatype.security.SecuritySystem;
-import org.sonatype.security.realms.XmlAuthenticatingRealm;
+import org.sonatype.security.authentication.AuthenticationException;
 
-public class MultipleRealmsTest
-    extends AbstractNexusLdapTestCase
+public class LdapNexusIT
+    extends NexusLdapTestSupport
 {
 
     @Test
@@ -35,11 +31,24 @@ public class MultipleRealmsTest
         SecuritySystem security = lookup( SecuritySystem.class );
         security.start();
 
-        security.authenticate( new UsernamePasswordToken( "cstamas", "cstamas123" ) );
+        Assert.assertNotNull( security.authenticate( new UsernamePasswordToken( "cstamas", "cstamas123" ) ) );
+    }
 
-        security.authenticate( new UsernamePasswordToken( "admin", "admin123" ) );
+    @Test
+    public void testAuthenticationFailure()
+        throws Exception
+    {
+        SecuritySystem security = lookup( SecuritySystem.class );
+        security.start();
 
-        security.authenticate( new UsernamePasswordToken( "deployment", "deployment123" ) );
+        try
+        {
+            Assert.assertNull( security.authenticate( new UsernamePasswordToken( "cstamas", "INVALID" ) ) );
+        }
+        catch ( AuthenticationException e )
+        {
+            // expected
+        }
     }
 
     @Test
@@ -49,23 +58,11 @@ public class MultipleRealmsTest
         SecuritySystem security = lookup( SecuritySystem.class );
         security.start();
 
-        // LDAP user
         SimplePrincipalCollection principals = new SimplePrincipalCollection();
         principals.add( "cstamas", new NexusLdapAuthenticationRealm().getName() );
 
-        Assert.assertTrue( security.hasRole( principals, "nx-developer" ) );
+        Assert.assertTrue( security.hasRole( principals, "developer" ) );
         Assert.assertFalse( security.hasRole( principals, "JUNK" ) );
-
-        // xml user
-        principals = new SimplePrincipalCollection();
-        // users must be from the correct realm now!
-        // TODO: bdemers or dbradicich, this "fix" is wrong, it relies on imple details!
-        // was: principals.add( "deployment", new XmlAuthenticatingRealm().getName() );
-        principals.add( "deployment", XmlAuthenticatingRealm.ROLE );
-
-        Assert.assertTrue( security.hasRole( principals, "nx-deployment" ) );
-        Assert.assertFalse( security.hasRole( principals, "JUNK" ) );
-
     }
 
     @Test
@@ -75,32 +72,10 @@ public class MultipleRealmsTest
         SecuritySystem security = lookup( SecuritySystem.class );
         security.start();
 
-        // LDAP
         SimplePrincipalCollection principals = new SimplePrincipalCollection();
         principals.add( "cstamas", new NexusLdapAuthenticationRealm().getName() );
 
         Assert.assertTrue( security.isPermitted( principals, "security:usersforgotpw:create" ) );
         Assert.assertFalse( security.isPermitted( principals, "security:usersforgotpw:delete" ) );
-
-        // XML
-        principals = new SimplePrincipalCollection();
-        // TODO: bdemers or dbradicich, this "fix" is wrong, it relies on imple details!
-        // was: principals.add( "test-user", new XmlAuthenticatingRealm().getName() );
-        principals.add( "test-user", XmlAuthenticatingRealm.ROLE );
-
-        Assert.assertTrue( security.isPermitted( principals, "security:usersforgotpw:create" ) );
-        Assert.assertFalse( security.isPermitted( principals, "security:usersforgotpw:delete" ) );
-
-        Assert.assertTrue( security.isPermitted( principals, "nexus:target:1:*:delete" ) );
-
     }
-
-    @Override
-    protected void copyDefaultConfigToPlace()
-        throws IOException
-    {
-        IOUtil.copy( getClass().getResourceAsStream( "/test-conf/security-configuration-multipleRealms.xml" ),
-                     new FileOutputStream( getSecurityConfiguration() ) );
-    }
-
 }
