+++ nexus/plugins/ldap/nexus-ldap-realm-plugin/src/test/java/org/sonatype/nexus/MultipleRealmsIT.java
@@ -12,21 +12,20 @@
  */
 package org.sonatype.nexus;
 
-import java.io.File;
+import java.io.FileOutputStream;
 import java.io.IOException;
 
 import org.apache.shiro.authc.UsernamePasswordToken;
 import org.apache.shiro.subject.SimplePrincipalCollection;
-import org.codehaus.plexus.context.Context;
+import org.codehaus.plexus.util.IOUtil;
 import org.junit.Assert;
 import org.junit.Test;
 import org.sonatype.nexus.security.ldap.realms.NexusLdapAuthenticationRealm;
 import org.sonatype.security.SecuritySystem;
-import org.sonatype.security.authentication.AuthenticationException;
 import org.sonatype.security.realms.XmlAuthenticatingRealm;
 
-public class MultipleRealmsLdapNotConfiguredTest
-    extends AbstractNexusLdapTestCase
+public class MultipleRealmsIT
+    extends NexusLdapTestSupport
 {
 
     @Test
@@ -36,48 +35,37 @@ public class MultipleRealmsLdapNotConfiguredTest
         SecuritySystem security = lookup( SecuritySystem.class );
         security.start();
 
-        // LDAP should fail
-        try
-        {
-            security.authenticate( new UsernamePasswordToken( "cstamas", "cstamas123" ) );
-            Assert.fail( "Expected AuthenticationException to be thrown." );
-        }
-        catch ( AuthenticationException e )
-        {
-            // expected
-        }
+        security.authenticate( new UsernamePasswordToken( "cstamas", "cstamas123" ) );
 
-        // xml should not
-        Assert.assertNotNull( security.authenticate( new UsernamePasswordToken( "admin", "admin123" ) ) );
+        security.authenticate( new UsernamePasswordToken( "admin", "admin123" ) );
 
-        Assert.assertNotNull( security.authenticate( new UsernamePasswordToken( "deployment", "deployment123" ) ) );
-
+        security.authenticate( new UsernamePasswordToken( "deployment", "deployment123" ) );
     }
 
     @Test
     public void testAuthorization()
         throws Exception
     {
-
         SecuritySystem security = lookup( SecuritySystem.class );
         security.start();
 
-        // LDAP should fail
+        // LDAP user
         SimplePrincipalCollection principals = new SimplePrincipalCollection();
         principals.add( "cstamas", new NexusLdapAuthenticationRealm().getName() );
 
-        // if realm is not configured, the user should not be able to be authorized
-        Assert.assertFalse( security.hasRole( principals, "nx-developer" ) );
+        Assert.assertTrue( security.hasRole( principals, "nx-developer" ) );
         Assert.assertFalse( security.hasRole( principals, "JUNK" ) );
 
         // xml user
         principals = new SimplePrincipalCollection();
+        // users must be from the correct realm now!
         // TODO: bdemers or dbradicich, this "fix" is wrong, it relies on imple details!
         // was: principals.add( "deployment", new XmlAuthenticatingRealm().getName() );
         principals.add( "deployment", XmlAuthenticatingRealm.ROLE );
 
         Assert.assertTrue( security.hasRole( principals, "nx-deployment" ) );
         Assert.assertFalse( security.hasRole( principals, "JUNK" ) );
+
     }
 
     @Test
@@ -91,8 +79,8 @@ public class MultipleRealmsLdapNotConfiguredTest
         SimplePrincipalCollection principals = new SimplePrincipalCollection();
         principals.add( "cstamas", new NexusLdapAuthenticationRealm().getName() );
 
-        // if realm is not configured, the user should not be able to be authorized
-        Assert.assertFalse( security.isPermitted( principals, "security:usersforgotpw:create" ) );
+        Assert.assertTrue( security.isPermitted( principals, "security:usersforgotpw:create" ) );
+        Assert.assertFalse( security.isPermitted( principals, "security:usersforgotpw:delete" ) );
 
         // XML
         principals = new SimplePrincipalCollection();
@@ -104,26 +92,15 @@ public class MultipleRealmsLdapNotConfiguredTest
         Assert.assertFalse( security.isPermitted( principals, "security:usersforgotpw:delete" ) );
 
         Assert.assertTrue( security.isPermitted( principals, "nexus:target:1:*:delete" ) );
+
     }
 
     @Override
     protected void copyDefaultConfigToPlace()
         throws IOException
     {
-        copyResource( "/test-conf/security-configuration-multipleRealms.xml", getSecurityConfiguration() );
-    }
-
-    @Override
-    protected void customizeContext( Context ctx )
-    {
-        super.customizeContext( ctx );
-
-        ctx.put( CONF_DIR_KEY, getLdapXml().getParentFile().getAbsolutePath() );
-    }
-
-    private File getLdapXml()
-    {
-        return new File( getConfHomeDir(), "no-conf/ldap.xml" );
+        IOUtil.copy( getClass().getResourceAsStream( "/test-conf/security-configuration-multipleRealms.xml" ),
+                     new FileOutputStream( getSecurityConfiguration() ) );
     }
 
 }
