+++ nexus/plugins/ldap/nexus-ldap-realm-plugin/src/test/java/org/sonatype/nexus/NexusLdapTestSupport.java
@@ -35,7 +35,7 @@ import org.sonatype.sisu.ehcache.CacheManagerComponent;
 
 import com.google.inject.Module;
 
-public abstract class AbstractNexusLdapTestCase
+public abstract class NexusLdapTestSupport
     extends NexusTestSupport
 {
     @Override
