+++ nexus/plugins/ldap/nexus-ldap-realm-plugin/src/test/java/org/sonatype/nexus/NotConfiguredLdapNexusIT.java
@@ -12,16 +12,19 @@
  */
 package org.sonatype.nexus;
 
+import java.io.File;
+
 import org.apache.shiro.authc.UsernamePasswordToken;
 import org.apache.shiro.subject.SimplePrincipalCollection;
+import org.codehaus.plexus.context.Context;
 import org.junit.Assert;
 import org.junit.Test;
-import org.sonatype.nexus.security.ldap.realms.NexusLdapAuthenticationRealm;
 import org.sonatype.security.SecuritySystem;
 import org.sonatype.security.authentication.AuthenticationException;
+import org.sonatype.security.ldap.realms.AbstractLdapAuthenticatingRealm;
 
-public class LdapNexusTest
-    extends AbstractNexusLdapTestCase
+public class NotConfiguredLdapNexusIT
+    extends NexusLdapTestSupport
 {
 
     @Test
@@ -31,19 +34,10 @@ public class LdapNexusTest
         SecuritySystem security = lookup( SecuritySystem.class );
         security.start();
 
-        Assert.assertNotNull( security.authenticate( new UsernamePasswordToken( "cstamas", "cstamas123" ) ) );
-    }
-
-    @Test
-    public void testAuthenticationFailure()
-        throws Exception
-    {
-        SecuritySystem security = lookup( SecuritySystem.class );
-        security.start();
-
         try
         {
-            Assert.assertNull( security.authenticate( new UsernamePasswordToken( "cstamas", "INVALID" ) ) );
+            security.authenticate( new UsernamePasswordToken( "cstamas", "cstamas123" ) );
+            Assert.fail( "Expected AuthenticationException to be thrown." );
         }
         catch ( AuthenticationException e )
         {
@@ -59,9 +53,11 @@ public class LdapNexusTest
         security.start();
 
         SimplePrincipalCollection principals = new SimplePrincipalCollection();
-        principals.add( "cstamas", new NexusLdapAuthenticationRealm().getName() );
+        principals.add( "cstamas", AbstractLdapAuthenticatingRealm.class.getName() );
 
-        Assert.assertTrue( security.hasRole( principals, "developer" ) );
+        // if realm is not configured, the user should not be able to be authorized
+
+        Assert.assertFalse( security.hasRole( principals, "developer" ) );
         Assert.assertFalse( security.hasRole( principals, "JUNK" ) );
     }
 
@@ -73,9 +69,23 @@ public class LdapNexusTest
         security.start();
 
         SimplePrincipalCollection principals = new SimplePrincipalCollection();
-        principals.add( "cstamas", new NexusLdapAuthenticationRealm().getName() );
+        principals.add( "cstamas", AbstractLdapAuthenticatingRealm.class.getName() );
 
-        Assert.assertTrue( security.isPermitted( principals, "security:usersforgotpw:create" ) );
-        Assert.assertFalse( security.isPermitted( principals, "security:usersforgotpw:delete" ) );
+        // if realm is not configured, the user should not be able to be authorized
+        Assert.assertFalse( security.isPermitted( principals, "nexus:usersforgotpw:create" ) );
     }
+
+    @Override
+    protected void customizeContext( Context ctx )
+    {
+        super.customizeContext( ctx );
+
+        ctx.put( CONF_DIR_KEY, getLdapXml().getParentFile().getAbsolutePath() );
+    }
+
+    private File getLdapXml()
+    {
+        return new File( getConfHomeDir(), "no-conf/ldap.xml" );
+    }
+
 }
