+++ nexus/plugins/ldap/nexus-ldap-realm-plugin/src/test/java/org/sonatype/nexus/security/ldap/realms/test/api/AuthMd5ConnectionIT.java
@@ -22,14 +22,14 @@ import org.codehaus.plexus.util.IOUtil;
 import org.restlet.data.Request;
 import org.restlet.data.Response;
 import org.restlet.resource.ResourceException;
-import org.sonatype.nexus.AbstractNexusLdapTestCase;
+import org.sonatype.nexus.NexusLdapTestSupport;
 import org.sonatype.nexus.security.ldap.realms.api.dto.LdapConnectionInfoDTO;
 import org.sonatype.nexus.security.ldap.realms.test.api.dto.LdapAuthenticationTestRequest;
 import org.sonatype.plexus.rest.resource.PlexusResource;
 
 
-public class AuthMd5ConnectionTest
-    extends AbstractNexusLdapTestCase
+public class AuthMd5ConnectionIT
+    extends NexusLdapTestSupport
 {
 
     private PlexusResource getResource()
