+++ nexus/plugins/ldap/nexus-ldap-realm-plugin/src/test/resources/org/sonatype/nexus/NotConfiguredLdapNexusIT.xml
