+++ nexus/plugins/ldap/nexus-ldap-realm-plugin/src/test/resources/org/sonatype/nexus/security/ldap/realms/test/api/UserGroupConfigIT.xml
